Extract concurrent identity hashmap into separate class
